requirements.txt
flask==3.1.0
openai==1.86.0
google-cloud-speech==2.30.0
pydub==0.25.1
python-dotenv==1.1.1
audioop-lts
